requirements.txt
Flask==3.0.3
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pyjwt==2.7.0
Flask-Cors==4.0.1